Move get-postback-or-throw under lib/postback/api
Co-locate the helper with other postback API modules and point partner-profile postback routes at the new path.
