+++ apps/web/app/(ee)/api/partner-profile/postbacks/[postbackId]/events/route.ts
@@ -1,6 +1,6 @@
-import { getPostbackOrThrow } from "@/lib/api/postbacks/get-postback-or-throw";
 import { withPartnerProfile } from "@/lib/auth/partner";
 import { getPostbackEvents } from "@/lib/postback/api/get-postback-events";
+import { getPostbackOrThrow } from "@/lib/postback/api/get-postback-or-throw";
 import { NextResponse } from "next/server";
 
 // GET /api/partner-profile/postbacks/[postbackId]/events
