+++ apps/web/app/(ee)/api/partner-profile/postbacks/[postbackId]/route.ts
@@ -1,6 +1,6 @@
-import { getPostbackOrThrow } from "@/lib/api/postbacks/get-postback-or-throw";
 import { parseRequestBody } from "@/lib/api/utils";
 import { withPartnerProfile } from "@/lib/auth/partner";
+import { getPostbackOrThrow } from "@/lib/postback/api/get-postback-or-throw";
 import {
   postbackSchema,
   updatePostbackInputSchema,
